Refactor: move ServerState to server_state.rs
