+++ openraft/src/core/mod.rs
@@ -12,6 +12,7 @@ mod raft_core;
 pub(crate) mod replication;
 #[cfg(test)]
 mod replication_state_test;
+mod server_state;
 mod vote;
 
 use std::fmt::Debug;
@@ -25,8 +26,7 @@ use leader_state::LeaderState;
 use learner_state::LearnerState;
 use rand::thread_rng;
 use rand::Rng;
-use serde::Deserialize;
-use serde::Serialize;
+pub use server_state::ServerState;
 use tokio::sync::broadcast;
 use tokio::sync::mpsc;
 use tokio::sync::oneshot;
@@ -603,51 +603,6 @@ pub(self) enum SnapshotUpdate<C: RaftTypeConfig> {
     SnapshotComplete(LogId<C::NodeId>),
     /// Snapshot creation failed.
     SnapshotFailed,
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////////////
-
-/// All possible states of a Raft node.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub enum ServerState {
-    /// The node is completely passive; replicating entries, but neither voting nor timing out.
-    Learner,
-    /// The node is replicating logs from the leader.
-    Follower,
-    /// The node is campaigning to become the cluster leader.
-    Candidate,
-    /// The node is the Raft cluster leader.
-    Leader,
-    /// The Raft node is shutting down.
-    Shutdown,
-}
-
-impl Default for ServerState {
-    fn default() -> Self {
-        Self::Follower
-    }
-}
-
-impl ServerState {
-    /// Check if currently in learner state.
-    pub fn is_learner(&self) -> bool {
-        matches!(self, Self::Learner)
-    }
-
-    /// Check if currently in follower state.
-    pub fn is_follower(&self) -> bool {
-        matches!(self, Self::Follower)
-    }
-
-    /// Check if currently in candidate state.
-    pub fn is_candidate(&self) -> bool {
-        matches!(self, Self::Candidate)
-    }
-
-    /// Check if currently in leader state.
-    pub fn is_leader(&self) -> bool {
-        matches!(self, Self::Leader)
-    }
 }
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
