+++ openraft/src/core/server_state.rs
@@ -1,0 +1,45 @@
+use serde::Deserialize;
+use serde::Serialize;
+
+/// All possible states of a Raft node.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum ServerState {
+    /// The node is completely passive; replicating entries, but neither voting nor timing out.
+    Learner,
+    /// The node is replicating logs from the leader.
+    Follower,
+    /// The node is campaigning to become the cluster leader.
+    Candidate,
+    /// The node is the Raft cluster leader.
+    Leader,
+    /// The Raft node is shutting down.
+    Shutdown,
+}
+
+impl Default for ServerState {
+    fn default() -> Self {
+        Self::Follower
+    }
+}
+
+impl ServerState {
+    /// Check if currently in learner state.
+    pub fn is_learner(&self) -> bool {
+        matches!(self, Self::Learner)
+    }
+
+    /// Check if currently in follower state.
+    pub fn is_follower(&self) -> bool {
+        matches!(self, Self::Follower)
+    }
+
+    /// Check if currently in candidate state.
+    pub fn is_candidate(&self) -> bool {
+        matches!(self, Self::Candidate)
+    }
+
+    /// Check if currently in leader state.
+    pub fn is_leader(&self) -> bool {
+        matches!(self, Self::Leader)
+    }
+}
